move helper/docoder to pkg/decoder
