+++ docs/cn/data-pipeline/extension/ext-default-decoder.md
@@ -12,7 +12,7 @@
 ## 配置参数
 | 参数                | 类型      | 是否必选 | 说明                                                                                                  |
 |-------------------|---------|------|-----------------------------------------------------------------------------------------------------|
-| Format            | String  | 是    | 具体的协议，[查看支持的具体协议列表](https://github.com/alibaba/ilogtail/blob/master/helper/decoder/common/comon.go) |
+| Format            | String  | 是    | 具体的协议，[查看支持的具体协议列表](https://github.com/alibaba/ilogtail/blob/master/pkg/protocol/decoder/common/comon.go) |
 | FieldsExtend      | Boolean | 否    | 是否启用增强字段功能，默认为false，仅针对Format=influxdb时有效                                                           |
 | DisableUncompress | Boolean | 是    | 否不解压数据，默认为false，仅针对Format=raw时有效                                                                    |
