+++ pkg/protocol/decoder/common/comon.go
