+++ pkg/protocol/decoder/influxdb/decoder_test.go
