+++ pkg/protocol/decoder/raw/decoder_test.go
