+++ plugins/input/opentelemetry/input_server.go
@@ -30,11 +30,11 @@ import (
 	"go.opentelemetry.io/collector/pdata/ptrace/ptraceotlp"
 	"google.golang.org/grpc"
 
-	"github.com/alibaba/ilogtail/helper/decoder/common"
-	"github.com/alibaba/ilogtail/helper/decoder/opentelemetry"
 	"github.com/alibaba/ilogtail/pkg/logger"
 	"github.com/alibaba/ilogtail/pkg/pipeline"
 	"github.com/alibaba/ilogtail/pkg/pipeline/extensions"
+	"github.com/alibaba/ilogtail/pkg/protocol/decoder/common"
+	"github.com/alibaba/ilogtail/pkg/protocol/decoder/opentelemetry"
 	"github.com/alibaba/ilogtail/plugins/input/httpserver"
 )
