+++ plugins/input/opentelemetry/receiver.go
@@ -25,8 +25,8 @@ import (
 	"go.opentelemetry.io/collector/pdata/ptrace"
 	"go.opentelemetry.io/collector/pdata/ptrace/ptraceotlp"
 
-	"github.com/alibaba/ilogtail/helper/decoder/opentelemetry"
 	"github.com/alibaba/ilogtail/pkg/pipeline"
+	"github.com/alibaba/ilogtail/pkg/protocol/decoder/opentelemetry"
 )
 
 // The following functions implement Opten telemetry GRPCServer Interface.
